v1.0 demo

+++ Project/Easy_Diet/app/src/main/java/au/edu/sydney/comp5216/easydiet/profile/HighProteinCalculator.java
@@ -18,7 +18,7 @@ public class HighProteinCalculator extends DietCalculator {
 
     public ArrayList<Double> calculateDailyProteinTarget(){
 
-        int weight = user.getWeight();
+        double weight = user.getWeight();
         ArrayList<Double> dailyProteinTargetRange = new ArrayList<>();
 
         double dailyProteinLower = 1.6 * weight;
